requirements.txt
pandas
transformers
sentence-transformers
faiss-cpu
torch
nltk
flask